% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugScreening-internal.R
\encoding{UTF-8}
\name{filterDrugScreenOneDrug}
\alias{filterDrugScreenOneDrug}
\title{Filter the drug screen dataset using specific information: drug name,
dosage type, name of the study and screen type.}
\usage{
filterDrugScreenOneDrug(drugData, drugName, study, screenType, doseType)
}
\arguments{
\item{drugData}{a \code{data.frame} containing the drug screening
information. The mandatory columns are: 'organoid_id', 'timestamp', 'study',
'screen_type', 'dosage_type', 'drug_a', 'drug_b', 'drug_c',
'drug_background',and 'relative_auc'.}

\item{drugName}{a single \code{character} string representing the
selected drug.}

\item{study}{a \code{vector} of \code{character} string representing the
name(s) of
the study selected for the analyses. The study must be present in the drug
screening dataset. The study can be found in the 'study' column of the
drug screening dataset.}

\item{screenType}{a \code{vector} of \code{character} string representing
the type of screening selected for the analyses. The type must be present
in the drug screening dataset. The screen type can be found in
the 'screen_type' column of the drug screening dataset.}

\item{doseType}{a single \code{character} string representing the
selected dosage type. Default: \code{"Averaged"}.}
}
\value{
a \code{data.frame} containing the selected organoids according to
the parameter specification. Those columns are present: 'organoid_id',
'timestamp', dosage_type', 'drug_a', 'drug_b', 'drug_c', 'drug_background',
and 'relative_auc'.
}
\description{
The function extracts information from a drug screen dataset
that is related to the specified parameters. The filters are specific to
the drug name ('drug_a' column), the dosage type ('dodsage_type' column),
the name of the study ('study' column) and
the screen type ('dosage_type' column).
}
\examples{

## Load drug screen dataset for 1 drug
data(simpleDrugScreening)

## Filter the drug screening dataset to retain specific rows
results <- OrganoidsToolBox:::filterDrugScreenOneDrug(
    drugData=simpleDrugScreening, drugName="Methotrexate",
    study="MEGA-TEST", screenType="TEST-01", doseType="Averaged")


}
\author{
Pascal Belleau, Astrid Deschênes
}
\keyword{internal}
