% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugScreening-internal.R
\encoding{UTF-8}
\name{findQuantileOneDrug}
\alias{findQuantileOneDrug}
\title{Extract the sensitive and resistant organoids
for a specific drug from a drug screening dataset}
\usage{
findQuantileOneDrug(cleanDrugData, quantile)
}
\arguments{
\item{cleanDrugData}{a \code{data.frame} containing the drug screening
information. The mandatory columns are: 'organoid_id', 'timestamp',
'dosage_type', 'drug_a', 'drug_b', 'drug_c', 'drug_background',
and 'relative_auc'.}

\item{quantile}{a single positive \code{numeric} between \code{0} and
\code{0.5} representing the cutoff for the organoids selection.}
}
\value{
a \code{list} containing 3 entries:
\itemize{
\item{quantile}{the \code{upper} and \code{lower} quantiles for the
specified drug.}
\item{dataset}{a \code{data.frame} containing the data used to
calculate the quantiles. }
\item{extreme}{a \code{data.frame} containing the sensitive and
resistant organoids according to the specified quantiles.
The \code{data.frame} also contains the relative AUC. }
}
}
\description{
The function extract the sensitive and resistant organoids
for a specific drug. The organoids with an average AUC equal or inferior to
the specified quantile are considered sensitive while the organoids with an
average AUC superior or equal to (1 - quantile) are considered resistant.
The function requires a minimum of 3 values to run the quantile analysis.
}
\examples{

## Load drug screen dataset for 1 drug
data(simpleDrugScreening)

## Filter the drug screen dataset
cleanData <- OrganoidsToolBox:::filterDrugScreenOneDrug(
    drugData=simpleDrugScreening, drugName="Methotrexate",
    study="MEGA-TEST", screenType="TEST-01", doseType="Averaged")

## Calculate the extreme organoids for the cleaned methotrexate drug
## screening using a quantile of 1/4
results <- OrganoidsToolBox:::findQuantileOneDrug(
    cleanDrugData=cleanData, quantile=1/4)

## The classification of the organoids is in the 'extreme' entry
results$extreme

}
\author{
Pascal Belleau, Astrid Deschênes
}
\keyword{internal}
