% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganoidsToolBox.R
\docType{package}
\encoding{UTF-8}
\name{OrganoidsToolBox-package}
\alias{OrganoidsToolBox-package}
\alias{OrganoidsToolBox}
\title{OrganoidsToolBox: TODO}
\description{
To learn more about \pkg{OrganoidsToolBox} package see:
\url{https://github.com/KrasnitzLab/OrganoidsToolBox/wiki}
}
\seealso{
\itemize{
\item \code{\link{getClassOneDrug}} {for selecting the organoids
with sensitive and resistant behavior for a specific drug screening.}
}
}
\author{
Maintainer:
Astrid Deschênes <adeschen@hotmail.com>
}
\keyword{package}
