% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/OrganoidsToolBox.R
\docType{data}
\name{patientInfo}
\alias{patientInfo}
\title{Simple demo patient information dataset. The information is related to the
organoids present in the 'drugScreening' dataset.}
\format{
a \code{data.frame} with the following columns:
\itemize{
\item \code{organoid_id} { a \code{character} string representing the
organoid identifier. }
\item \code{patient_id} { a \code{character} string representing the patient
identifier. }
}
}
\usage{
data(drugScreening)
}
\description{
Simple demo patient information dataset. The information is related to the
organoids present in the 'drugScreening' dataset.
}
\examples{

## Load drug screen dataset for 1 drug
data(drugScreening)

## Load patient information dataset for 1 drug
data(patientInfo)

## Select unrelated organoids in the drug screening
results <- selectNoReplicateOrganoids(drugScreening=drugScreening,
    patientInfo=patientInfo)

## The drug screen dataset has been filtered
results


}
\seealso{
\itemize{
\item \code{\link{getClassOneDrug}} { for selecting the samples
with sensitive and resistant behavior for a
specific drug screening dataset.}
\item \code{\link{selectNoReplicateOrganoids}} { for selecting
the unrelated samples (only one sample per patient) from a drug screening
dataset.}
}
}
\keyword{datasets}
