% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugScreening.R
\encoding{UTF-8}
\name{plotDrugAUCDistribution}
\alias{plotDrugAUCDistribution}
\title{TODO}
\usage{
plotDrugAUCDistribution(drugQuantile)
}
\arguments{
\item{drugQuantile}{TODO}
}
\value{
TODO
}
\description{
The function TODO
}
\examples{

## TODO
drugName <- "Methotrexate"

}
\author{
Astrid Deschênes, Pascal Belleau
}
