% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugScreening.R
\encoding{UTF-8}
\name{selectNoReplicateOrganoids}
\alias{selectNoReplicateOrganoids}
\title{Select the TODO}
\usage{
selectNoReplicateOrganoids(drugScreening, patientInfo)
}
\arguments{
\item{drugScreening}{a \code{data.frame} that contains the drug screening
results. The column is 'organoid_id' is mandatory.}

\item{patientInfo}{a \code{data.frame} containing the meta-data information
related to the organoids. The mandatory columns are: 'organoid_id'
and 'patient_id'.}
}
\value{
a \code{data.frame} with the drug screening results filtered
to contain with only one organoid per patient. The extra columns from the
patient information dataset is added as extra columns to the final dataset.
}
\description{
The function TODO
}
\examples{

## Load drug screen dataset for 1 drug
data(drugScreening)

## Load patient information dataset for 1 drug
data(patientInfo)

## Set seed to get reproducible results
set.seed(1212)

results <- selectNoReplicateOrganoids(drugScreening=drugScreening,
    patientInfo=patientInfo)

## TODO
results

}
\author{
Astrid Deschênes, Pascal Belleau
}
