% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugScreening.R
\encoding{UTF-8}
\name{selectOrgForOneDrug}
\alias{selectOrgForOneDrug}
\title{Select the organoids with sensitive and resistant behavior for a
specific drug screening}
\usage{
selectOrgForOneDrug(
  drugScreening,
  drugName,
  study,
  screenType,
  doseType = "Averaged",
  quantile = 1/3
)
}
\arguments{
\item{drugScreening}{a single \code{character} string representing the path
and name of the drug screening file that contains the information needed
to run the organoid selection. Those columns are mandatory: 'organoid_id',
'timestamp', 'study', 'screen_type', 'dosage_type',
'drug_a', 'drug_b', 'drug_c', 'drug_background' and 'relative_auc'.}

\item{drugName}{a single \code{character} string representing the name of
the drug selected for the analyses. The drug must be present in the drug
screening dataset. The drug name can be found in the 'drug_a' column of the
drug screening dataset.}

\item{study}{a single \code{character} string representing the name of
the study selected for the analyses. The study must be present in the drug
screening dataset. The study can be found in the 'study' column of the
drug screening dataset.}

\item{screenType}{a \code{vector} of \code{character} string representing
the type of
screening selected for the analyses. The type must be present in the drug
screening dataset.  The screenType can be found in the 'screen_type'
column of the
drug screening dataset.}

\item{doseType}{a single \code{character} string representing the type of
dosage selected for the analyses. The type must be present in the drug
screening dataset.}

\item{quantile}{a single positive \code{numeric} between 0 and 0.5
indicating the quantile used to select the organoids. Default: \code{1/3}.}
}
\value{
an object of class "\code{DrugAUCQuantile}" which contains the
sensitive and resistant organoids for a specific drug. This object is a
\code{list} with the following 3 components:
\itemize{
\item{quantile}{the \code{upper} and \code{lower} quantiles for the
specified drug.}
\item{dataset}{a \code{data.frame} containing the data used to
calculate the quantiles. }
\item{extreme}{a \code{data.frame} containing the sensitive and
resistant organoids according to the specified quantiles.
The \code{data.frame} also contains the relative AUC. }
}
}
\description{
The function selects the organoids within the low and upper
quantile, as specified by user, for a specific drug screening.
}
\examples{

## TODO
drugName <- "Methotrexate"

}
\author{
Astrid Deschênes, Pascal Belleau
}
