% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugScreening-internal.R
\encoding{UTF-8}
\name{validateSelectOrgWithoutReplicateForOneDrug}
\alias{validateSelectOrgWithoutReplicateForOneDrug}
\title{Validate input parameters for the
getClassNoReplicateOneDrug() function}
\usage{
validateSelectOrgWithoutReplicateForOneDrug(
  drugScreening,
  drugName,
  study,
  screenType,
  patientInfo,
  doseType,
  quantile
)
}
\arguments{
\item{drugScreening}{a single \code{character} string representing the path
and name of the drug screening file that contains the information needed
to run the organoid selection. Those columns are mandatory: 'organoid_id',
'timestamp', 'study', 'screen_type', 'dosage_type',
'drug_a', 'drug_b', 'drug_c', 'drug_background' and 'relative_auc'.}

\item{drugName}{a single \code{character} string representing the name of
the drug selected for the analyses.}

\item{study}{a single \code{character} string representing the name of
the study selected for the analyses.}

\item{screenType}{a \code{vector} of \code{character} string representing
the type of
screening selected for the analyses.}

\item{patientInfo}{a \code{data.frame} containing the meta-data information
related to the organoids. The mandatory columns are: 'organoid_id'
and 'patient_id'.}

\item{doseType}{a single \code{character} string representing the type of
dosage selected for the analyses.}

\item{quantile}{a single positive \code{numeric} between 0 and 0.5
indicating the quantile used to select the organoids. Default: \code{1/3}.}
}
\value{
the value \code{0L} when successful.
}
\description{
The function validates the input parameters for the
getClassNoReplicateOneDrug() function.
}
\examples{

## Load drug screen dataset for 1 drug
data(drugScreening)

## Load patient information dataset
data(patientInfo)

## Validation should return OL as all parameters are valids
OrganoidsToolBox:::validateSelectOrgWithoutReplicateForOneDrug(
    drugScreening=drugScreening,
    drugName="Methotrexate", study="MEGA-TEST", screenType="TEST-01",
    patientInfo=patientInfo, doseType="Averaged", quantile=1/3)

}
\author{
Astrid Deschênes, Pascal Belleau
}
\keyword{internal}
