% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/enrichment.R
\encoding{UTF-8}
\name{fisherCategoricalVariable}
\alias{fisherCategoricalVariable}
\title{Run Fisher tests in the sensitive and resistant groups
using a categorical value}
\usage{
fisherCategoricalVariable(drugQuantile, patientInfo, category)
}
\arguments{
\item{drugQuantile}{an object of class "\code{DrugAUCQuantile}" which
contains the sensitive and resistant organoids for a specific drug.}

\item{patientInfo}{a \code{data.frame} containing the meta-data information
related to the organoids. The mandatory columns are: 'organoid_id'
and 'patient_id'.}

\item{category}{a single \code{character} string representing the column
from the 'patientInfo' dataset that should be used as the
categorical variable.}
}
\value{
a \code{list} containing 2 components:
\itemize{
\item{Fisher}{ TODO}
\item{table}{a \code{data.frame} TODO }
}
}
\description{
The function run Fisher test in the sensitive and
resistant groups using a categorical value. TODO
}
\examples{

## Load drug screen dataset for 1 drug
data(drugScreening)

## Calculate the extreme organoids for the methotrexate drug screening
## using a quantile of 1/3
results <- getClassOneDrug(drugScreening=drugScreening,
    drugName="Methotrexate", study="MEGA-TEST", screenType="TEST-01",
    doseType="Averaged", quantile=1/3)

## The classification of the organoids is in the 'extreme' entry
results$extreme

}
\author{
Astrid Deschênes, Pascal Belleau
}
