% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugScreening.R
\encoding{UTF-8}
\name{plotDrugAUCDensityCurve}
\alias{plotDrugAUCDensityCurve}
\title{TODO}
\usage{
plotDrugAUCDensityCurve(drugQuantile)
}
\arguments{
\item{drugQuantile}{an object of class "\code{DrugAUCQuantile}" which
contains the
sensitive and resistant organoids for a specific drug.}
}
\value{
TODO
}
\description{
The function TODO
}
\examples{

## TODO
drugName <- "Methotrexate"

}
\author{
Astrid Deschênes, Pascal Belleau
}
