% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugScreening.R
\encoding{UTF-8}
\name{plotDrugAUCDensityCurve}
\alias{plotDrugAUCDensityCurve}
\title{Create a density plot using the relative AUC and the sensitivy
classes (SENSITIVE, AVERAGE, RESISTANT) of the organoids}
\usage{
plotDrugAUCDensityCurve(drugQuantile, byGroup = FALSE)
}
\arguments{
\item{drugQuantile}{an object of class "\code{DrugAUCQuantile}" which
contains the
sensitive and resistant organoids for a specific drug.}

\item{byGroup}{a \code{logical} indicating if the density is split
by group. Default: \code{FALSE}.}
}
\value{
a \code{ggplot} object that represents a density plot of the
AUC according to the drug senstivity classification(SENSITIVE, AVERAGE,
RESISTANT).
}
\description{
The function TODO
}
\examples{

## Load drug screen dataset for 1 drug
data(drugScreening)

## Calculate the extreme organoids for the methotrexate drug screening
## using a quantile of 1/3
results <- selectOrgForOneDrug(drugScreening=drugScreening,
    drugName="Methotrexate", study="MEGA-TEST", screenType="TEST-01",
    doseType="Averaged", quantile=1/3)

## Create a density plot where drug sensitivy groups are split
densityGraph <- plotDrugAUCDensityCurve(results, byGroup=TRUE)
densityGraph

}
\author{
Astrid Deschênes, Pascal Belleau
}
