% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugScreening.R
\encoding{UTF-8}
\name{plotDrugAUCViolinPlot}
\alias{plotDrugAUCViolinPlot}
\title{TODO}
\usage{
plotDrugAUCViolinPlot(drugQuantile, min = 0, max = 100, trim = FALSE)
}
\arguments{
\item{drugQuantile}{an object of class "\code{DrugAUCQuantile}" which
contains the
sensitive and resistant organoids for a specific drug.}

\item{min}{a single \code{numeric} representing the minimum value of the
y-axis. Default: \code{0}.}

\item{max}{a single \code{numeric} representing the maximum value of the
y-axis. The maximum value must be superior to the mimimum value.
Default: \code{100}.}

\item{trim}{a \code{logical} indicating if the tails of the violins
are trimmed to the range of the data. Default: \code{FALSE}.}
}
\value{
a \code{ggplot} object for a violin plot built with the specified
data.
}
\description{
The function TODO
}
\examples{

## Load drug screen dataset for 1 drug
data(drugScreening)

## Calculate the extreme organoids for the methotrexate drug screening
## using a quantile of 1/3
results <- selectOrgForOneDrug(drugScreening=drugScreening,
    drugName="Methotrexate", study="MEGA-TEST", screenType="TEST-01",
    doseType="Averaged", quantile=1/3)

## Plot results
p <- plotDrugAUCViolinPlot(drugQuantile=results, min=20, max=80, trim=TRUE)
plot(p)

}
\author{
Astrid Deschênes, Pascal Belleau
}
