% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugScreening.R
\encoding{UTF-8}
\name{selectOrgForOneDrug}
\alias{selectOrgForOneDrug}
\title{Select the organoids with sensitive and resistant behavior for a
specific drug screening}
\usage{
selectOrgForOneDrug(
  drugScreening,
  drugName,
  study,
  type,
  ancestry = c("ALL", "AFR", "EUR", "AMR", "EAS", "SAS"),
  quantile = 1/3
)
}
\arguments{
\item{drugScreening}{a single \code{character} string representing the path
and name of the drug screening file that contains the information needed
to run the organoid selection.}

\item{drugName}{a single \code{character} string representing the name of
the drug selected for the analyses. The drug must be present in the drug
screening file.}

\item{study}{a single \code{character} string representing the name of
the study selected for the analyses. The study must be present in the drug
screening file.}

\item{type}{a single \code{character} string representing the type of
study selected for the analyses. The type must be present in the drug
screening file.}

\item{ancestry}{a TODO. If `ancestry` is \code{"ALL"}, this
field is not used.}

\item{quantile}{a single positive \code{numeric} between 0 and 0.5
indicating the quantile used to select the organoids. Default: \code{1/3}.}
}
\value{
a TODO
}
\description{
The function selects the organoids within the low and upper
quantile, as specified by user, for a specific drug screening.
}
\examples{

## TODO
drugName <- "Methotrexate"

}
\author{
Astrid Deschênes, Pascal Belleau
}
